[add] 順番の入れ替え実装

+++ lib/phx_todo_app/todolist/item.ex
@@ -5,7 +5,7 @@ defmodule PhxTodoApp.Todolist.Item do
   schema "items" do
     field :done, :boolean, default: false
     field :title, :string
-    field :order, :integer, default: 1
+    field :order, :float, default: 1.0
 
     timestamps()
   end
